requirements.txt
# AI Avatar - Dependencies
# Python 3.10+

# Web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
jinja2>=3.1.0

# LLM - Google Gemini
google-generativeai==0.8.3

# TTS - ElevenLabs
elevenlabs==1.2.0

# Environment & config
python-dotenv==1.0.1

# Language detection (CZ/EN)
langdetect==1.0.9

# HTTP client
httpx==0.26.0
